Find Multi-Modal Nearest Neighbors

A wrapper function that finds Multi-Modal weights and generate a Multi-Modal NN object.

FindmmNN(
  object,
  sketch.assay = NULL,
  reduction.list,
  dims.list,
  k.nn = 20,
  knn.range = 200,
  l2.norm = TRUE,
  fix.wnn.weights = NULL,
  weighted.nn.name = "weighted.nn",
  verbose = TRUE,
  ...
)

Arguments

object

A Seurat object

sketch.assay

The name of the sketched assay of the landmark cells.

reduction.list

A list of dimensional reductions, one for each modality

dims.list

A list containing the dimensions for each reduction to use

k.nn

The number of nearest neighbors to compute for each modality. Internally clamped to a minimum of 20 (i.e. values below 20 are silently raised to 20).

knn.range

Range parameter for nearest neighbor search

l2.norm

Perform L2 normalization on the cell embeddings

fix.wnn.weights

Pre-specified modality weights. If provided, skips the calculation and uses these weights directly. Should be a list with the same length as reduction.list.

weighted.nn.name

Multimodal neighbor object name

verbose

Print progress bars and output

...

Arguments passed to other methods

Value

return a Seurat object that contains a weighted.nn Neighbor object between the landmark cells and all the other cells.

Details

This function is essentially a wrapper function that perform the following 2 procedures:

Seurat:::FindModalityWeights

Seurat:::MultiModalNN