Run Partial Least Squares (PLS) on Seurat Objects

Performs Partial Least Squares regression analysis on single-cell data. This function provides methods for plsr, spls, and cppls from the pls and spls packages.

Usage

RunPLS(object, ...)

# Default S3 method
RunPLS(
  object,
  assay = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

# S3 method for class 'Assay'
RunPLS(
  object,
  assay = NULL,
  features = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

# S3 method for class 'StdAssay'
RunPLS(
  object,
  assay = NULL,
  features = NULL,
  layer = "scale.data",
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

# S3 method for class 'Seurat'
RunPLS(
  object,
  assay = NULL,
  features = NULL,
  ncomp = 20,
  Y = NULL,
  Y.add = NULL,
  pls.function = c("plsr", "spls", "cppls"),
  verbose = TRUE,
  ndims.print = 1:5,
  nfeatures.print = 30,
  reduction.name = "pls",
  reduction.key = "PLS_",
  seed.use = 42,
  eta = 0.5,
  ...
)

Arguments

object

An object to run PLS on

...

Additional arguments to be passed to the PLS function

assay

Name of Assay PLS is being run on

ncomp

Number of components to compute

Y

a vector or matrix of responses, i.e., the dependent variable that PLS regresses X on. The length / number of rows should be the same as the number of cells. Y can have multiple columns.

Y.add

a vector or matrix of additional responses containing relevant information about the observations. Only used for cppls.

pls.function

PLS function from pls package to run (options: plsr, spls, cppls)

verbose

Print the top genes associated with high/low loadings for the components

ndims.print

components to print genes for

nfeatures.print

Number of genes to print for each component

reduction.name

the name of the DimReduc object

reduction.key

dimensional reduction key, specifies the string before the number for the dimension names.

seed.use

Set a random seed. Setting NULL will not set a seed.

eta

Thresholding parameter that controls the sparsity of the spls method (larger –> sparser). eta should be between 0 and 1.

features

Features to compute PLS on

layer

The layer in assay to use when running PLS analysis.

Value

Returns a DimReduc object with PLS results

Examples

if (FALSE) { # \dontrun{
# Run PLS on a Seurat object
seurat_obj <- RunPLS(seurat_obj, Y = "condition", ncomp = 10)
} # }